Remove JVM_OPTS export at top of init script
The current Cassandra setup causes JVM options to be passed twice when
starting Cassandra. This means that Java agents that don't gracefully
handle being loaded twice (e.g. the Prometheus JMX exporter) will cause
Cassandra to be unable to start up.

Removing this line fixes this problem. This commit is tagged at
1.24.0-monzo so we can pull it in to our current Puppet config. We will
be upgrading this module in due course (or moving off it), this is an
interim fix to get the JMX exporter up and running for now