Allow setting heapdump directory
The Debian Cassandra package includes an init script which hardcodes the
heapdump location to be within the Cassandra home directory, regardless
of what is specified in `-XX:HeapDumpPath`.

We need to move our heapdumps into the ephemeral disk so we don't run
out of space on our root EBS volumes.

Luckily, the Cassandra Puppet module already patches the default init
script to work around another Cassandra packaging bug, so we can easily
patch the patch.

Our patch removes the hard coding of this path, allowing the option to
be set using the `CASSANDRA_HEAPDUMP_DIR` environment variable, which is
then used by `cassandra-env.sh` to set the `-XX:HeapDumpPath` option.