Make the permissions cache max size configurable
We will do C* credential rotation through Vault and we will have a
credential pair per pod with permissions on the service keyspace.

If permissions are not stored in the cache, queries need to check
the system_auth keyspace (overhead of an extra query). We want to
allow configuration of the cache size to ensure it can hold all
permissions for all credentials, to avoid churn and overhead.